requirements.txt
torch
torchvision
basicsr
facexlib
gfpgan
realesrgan

streamlit 
opencv-python-headless
pillow